+++ packages/backend-plugin-api/src/testUtils.ts
@@ -14,8 +14,6 @@
  * limitations under the License.
  */
 
-// TODO: Remove this relative import when extrating the path utilities to this package
-// eslint-disable-next-line @backstage/no-relative-monorepo-imports
 import { packagePathMocks } from './paths';
 import { posix as posixPath, resolve as resolvePath } from 'path';
